chore: fix comments fix: get correct metadata version
