--- src/renderer/entities/metadata/index.ts
@@ -1,3 +1,0 @@
-export * from './lib';
-export * from './model/types';
-export * as metadataModel from './model/metadata';
