--- src/renderer/entities/metadata/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './metadataService';
-export * from './common/types';
