--- src/renderer/entities/metadata/model/types.ts
@@ -1,7 +1,0 @@
-import { ChainId, HexString } from '@renderer/domain/shared-kernel';
-
-export type Metadata = {
-  chainId: ChainId;
-  version: number;
-  metadata: HexString;
-};
